+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/OidcSiopVerifier.kt
@@ -91,13 +91,13 @@ class OidcSiopVerifier(
     }
 
     private val containerJwt =
-        FormatContainerJwt(algorithms = verifierJwsService.supportedAlgorithms.map { it.identifier }.toTypedArray())
+        FormatContainerJwt(algorithms = verifierJwsService.supportedAlgorithms.map { it.identifier })
 
     private val metadata by lazy {
         RelyingPartyMetadata(
-            redirectUris = arrayOf(relyingPartyUrl),
-            jsonWebKeySet = JsonWebKeySet(arrayOf(agentPublicKey.toJsonWebKey())),
-            subjectSyntaxTypesSupported = arrayOf(URN_TYPE_JWK_THUMBPRINT, PREFIX_DID_KEY),
+            redirectUris = listOf(relyingPartyUrl),
+            jsonWebKeySet = JsonWebKeySet(listOf(agentPublicKey.toJsonWebKey())),
+            subjectSyntaxTypesSupported = listOf(URN_TYPE_JWK_THUMBPRINT, PREFIX_DID_KEY),
             vpFormats = FormatHolder(
                 msoMdoc = containerJwt,
                 jwtVp = containerJwt,
@@ -250,8 +250,8 @@ class OidcSiopVerifier(
                 ConstantIndex.CredentialRepresentation.ISO_MDOC -> it.isoConstraint()
             }
         }
-        val attributeConstraint = requestedAttributes?.let { createConstraints(representation, it) } ?: arrayOf()
-        val constraintFields = listOfNotNull(typeConstraint, *attributeConstraint).toTypedArray()
+        val attributeConstraint = requestedAttributes?.let { createConstraints(representation, it) } ?: listOf()
+        val constraintFields = attributeConstraint + typeConstraint
         return AuthenticationRequestParameters(
             responseType = "$ID_TOKEN $VP_TOKEN",
             clientId = relyingPartyUrl,
@@ -266,11 +266,11 @@ class OidcSiopVerifier(
             presentationDefinition = PresentationDefinition(
                 id = uuid4().toString(),
                 formats = representation.toFormatHolder(),
-                inputDescriptors = arrayOf(
+                inputDescriptors = listOf(
                     InputDescriptor(
                         id = uuid4().toString(),
-                        schema = arrayOf(SchemaReference(credentialScheme?.schemaUri ?: "https://example.com")),
-                        constraints = Constraint(fields = constraintFields),
+                        schema = listOf(SchemaReference(credentialScheme?.schemaUri ?: "https://example.com")),
+                        constraints = Constraint(fields = constraintFields.filterNotNull()),
                     )
                 ),
             ),
@@ -284,7 +284,7 @@ class OidcSiopVerifier(
     }
 
     private fun ConstantIndex.CredentialScheme.vcConstraint() = ConstraintField(
-        path = arrayOf("$.type"),
+        path = listOf("$.type"),
         filter = ConstraintFilter(
             type = "string",
             pattern = vcType,
@@ -292,7 +292,7 @@ class OidcSiopVerifier(
     )
 
     private fun ConstantIndex.CredentialScheme.isoConstraint() = ConstraintField(
-        path = arrayOf("$.mdoc.doctype"),
+        path = listOf("$.mdoc.doctype"),
         filter = ConstraintFilter(
             type = "string",
             pattern = isoDocType,
@@ -302,12 +302,12 @@ class OidcSiopVerifier(
     private fun createConstraints(
         credentialRepresentation: ConstantIndex.CredentialRepresentation,
         attributeTypes: List<String>,
-    ): Array<ConstraintField> = attributeTypes.map {
+    ): Collection<ConstraintField> = attributeTypes.map {
         if (credentialRepresentation == ConstantIndex.CredentialRepresentation.ISO_MDOC)
-            ConstraintField(path = arrayOf("\$.mdoc.$it"), intentToRetain = false)
+            ConstraintField(path = listOf("\$.mdoc.$it"), intentToRetain = false)
         else
-            ConstraintField(path = arrayOf("\$.$it"))
-    }.toTypedArray()
+            ConstraintField(path = listOf("\$.$it"))
+    }
 
 
     sealed class AuthnResponseResult {
@@ -410,7 +410,7 @@ class OidcSiopVerifier(
         val presentationSubmission = params.presentationSubmission
             ?: return AuthnResponseResult.ValidationError("presentation_submission", params.state)
                 .also { Napier.w("presentation_submission empty") }
-        val descriptor = presentationSubmission.descriptorMap?.get(0)
+        val descriptor = presentationSubmission.descriptorMap?.firstOrNull()
             ?: return AuthnResponseResult.ValidationError("presentation_submission", params.state)
                 .also { Napier.w("presentation_submission contains no descriptors") }
         val vp = params.vpToken