+++ vclib/src/commonMain/kotlin/at/asitplus/wallet/lib/data/VerifiableCredential.kt
@@ -17,7 +17,7 @@ data class VerifiableCredential(
     @SerialName("id")
     val id: String,
     @SerialName("type")
-    val type: Array<String>,
+    val type: Collection<String>,
     @SerialName("issuer")
     val issuer: String,
     @Serializable(with = InstantStringSerializer::class)
@@ -43,7 +43,7 @@ data class VerifiableCredential(
         expirationDate: Instant? = Clock.System.now() + lifetime,
     ) : this(
         id = id,
-        type = arrayOf(VERIFIABLE_CREDENTIAL, credentialType),
+        type = listOf(VERIFIABLE_CREDENTIAL, credentialType),
         issuer = issuer,
         issuanceDate = issuanceDate,
         expirationDate = expirationDate,
@@ -61,7 +61,7 @@ data class VerifiableCredential(
         credentialType: String,
     ) : this(
         id = id,
-        type = arrayOf(VERIFIABLE_CREDENTIAL, credentialType),
+        type = listOf(VERIFIABLE_CREDENTIAL, credentialType),
         issuer = issuer,
         issuanceDate = issuanceDate,
         expirationDate = expirationDate,
@@ -77,37 +77,11 @@ data class VerifiableCredential(
         credentialSubject: RevocationListSubject,
     ) : this(
         id = id,
-        type = arrayOf(VERIFIABLE_CREDENTIAL, REVOCATION_LIST_2020),
+        type = listOf(VERIFIABLE_CREDENTIAL, REVOCATION_LIST_2020),
         issuer = issuer,
         issuanceDate = issuanceDate,
         expirationDate = issuanceDate + lifetime,
         credentialStatus = null,
         credentialSubject = credentialSubject,
     )
-
-    override fun equals(other: Any?): Boolean {
-        if (this === other) return true
-        if (other == null || this::class != other::class) return false
-
-        other as VerifiableCredential
-
-        if (id != other.id) return false
-        if (!type.contentEquals(other.type)) return false
-        if (issuer != other.issuer) return false
-        if (issuanceDate != other.issuanceDate) return false
-        if (expirationDate != other.expirationDate) return false
-        if (credentialStatus != other.credentialStatus) return false
-        return credentialSubject == other.credentialSubject
-    }
-
-    override fun hashCode(): Int {
-        var result = id.hashCode()
-        result = 31 * result + type.contentHashCode()
-        result = 31 * result + issuer.hashCode()
-        result = 31 * result + issuanceDate.hashCode()
-        result = 31 * result + (expirationDate?.hashCode() ?: 0)
-        result = 31 * result + (credentialStatus?.hashCode() ?: 0)
-        result = 31 * result + credentialSubject.hashCode()
-        return result
-    }
 }