+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release NEXT:
     - Given that all EC keys were previously uncompressed, different mutlicodec identifiers are now supported and the old encoding of uncompressed keys does not work anymore, as it was faulty.
     - In addition, the encoding of the mutlibase prefix has changed, since varint-Encoding is now used correctly. 
 - Fix name shadowing of gradle plugins by renaming file `Plugin.kt` -> `VcLibConventions.kt`
+ - Get rid of arrays in serializable types and use collections instead
 
 Release 3.4.0:
  - Target Java 17